Consolidate smart shape stuff
